pyproject.toml
[tool.poetry]
name = "telemerecsys"
version = "0.1.0"
description = "Teleme Recommendation Engine"
authors = ["Huvinesh Rajendran <huvinesh@teleme.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
qdrant-client = "^1.8.0"
sentence-transformers = "^2.5.1"
uvicorn = "^0.28.0"
fastapi = "^0.110.0"
pandas = "^2.2.1"
jax = "^0.4.25"
neo4j = "^5.18.0"
black = "^24.3.0"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
psycopg-binary = "^3.1.18"
psycopg = "^3.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
